Tile decorations generation using noise
There is of stuff to be commented here and I will comment it when I revisit the code later on. The tile probability was the hardest to implement since it has a lot to do with normal distribution p-value calculation (which I knew nothing about). If anyone has any suggestions to the probability calculation I created please share them with me.